pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = "AG-UI Chat Application with FastAPI and React"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "ag-ui-protocol>=0.0.41",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
]
